mito-ai: create tool executor
